trips-manager/requirements.txt
pymongo==4.6.1
kafka-python==2.0.2
requests==2.31.0